+++ Fleece/Support/sliceIO.cc
@@ -22,6 +22,7 @@
 
 #include "FleeceException.hh"
 #include "PlatformCompat.hh"
+#include "NumConversion.hh"
 #include <fcntl.h>
 #include <errno.h>
 
@@ -52,9 +53,9 @@ namespace fleece {
         fstat(fd, &stat);
         if (stat.st_size > SIZE_MAX)
             throw std::logic_error("File too big for address space");
-        alloc_slice data((size_t)stat.st_size);
-        ssize_t bytesRead = ::_read(fd, (void*)data.buf, data.size);
-        if (bytesRead < (ssize_t)data.size)
+        alloc_slice data(narrow_cast<size_t>(stat.st_size));
+        ssize_t bytesRead = narrow_cast<ssize_t>(::_read(fd, (void*)data.buf, narrow_cast<unsigned int>(data.size)));
+        if (bytesRead < narrow_cast<ssize_t>(data.size))
             FleeceException::_throwErrno("Can't read file %s", path);
         ::_close(fd);
         return data;
@@ -64,8 +65,8 @@ namespace fleece {
         int fd = ::_open(path, mode | O_WRONLY | O_BINARY, 0600);
         if (fd < 0)
             FleeceException::_throwErrno("Can't open file");
-        ssize_t written = ::_write(fd, s.buf, s.size);
-        if(written < (ssize_t)s.size)
+        ssize_t written = narrow_cast<ssize_t>(::_write(fd, s.buf, narrow_cast<unsigned int>(s.size)));
+        if(written < narrow_cast<ssize_t>(s.size))
             FleeceException::_throwErrno("Can't write file");
         ::_close(fd);
     }
